LLVM.General.AST

Description

This module and descendants define AST data types to represent LLVM code. Note that these types are designed for fidelity rather than convenience - if the truth of what LLVM supports is less than pretty, so be it.

Documentation

data Module

http://llvm.org/docs/LangRef.html#modulestructure

Constructors

Module

Fields

moduleName :: String

moduleDataLayout :: Maybe DataLayout

a DataLayout, if specified, must match that of the eventual code generator

moduleTargetTriple :: Maybe String

moduleDefinitions :: [Definition]

defaultModule :: Module

helper for making Modules

data Definition

Any thing which can be at the top level of a Module

Constructors

GlobalDefinition Global

TypeDefinition Name (Maybe Type)

MetadataNodeDefinition MetadataNodeID [Maybe Operand]

NamedMetadataDefinition String [MetadataNodeID]

ModuleInlineAssembly String

data Global

http://llvm.org/doxygen/classllvm_1_1GlobalValue.html

Constructors

GlobalVariable Name Linkage Visibility Bool AddrSpace Bool Bool Type (Maybe Constant) (Maybe String) Word32
http://llvm.org/docs/LangRef.html#global-variables

GlobalAlias Name Linkage Visibility Type Constant
http://llvm.org/docs/LangRef.html#aliases

Function Linkage Visibility CallingConvention [ParameterAttribute] Type Name ([Parameter], Bool) [FunctionAttribute] (Maybe String) Word32 [BasicBlock]
http://llvm.org/docs/LangRef.html#functions

globalVariableDefaults :: Global

helper for making GlobalVariables

globalAliasDefaults :: Global

helper for making GlobalAliass

functionDefaults :: Global

helper for making Functions

data Parameter

Parameters for Functions

Constructors

Parameter Type Name [ParameterAttribute]

data BasicBlock

http://llvm.org/doxygen/classllvm_1_1BasicBlock.html LLVM code in a function is a sequence of BasicBlocks each with a label, some instructions, and a terminator.

Constructors

BasicBlock Name [Named Instruction] (Named Terminator)

module LLVM.General.AST.Instruction

module LLVM.General.AST.Name

module LLVM.General.AST.Operand

module LLVM.General.AST.Type