LLVM.General.AST.DataLayout

Description

http://llvm.org/docs/LangRef.html#data-layout

Documentation

data Endianness

Little Endian is the one true way :-). Sadly, we must support the infidels.

Constructors

LittleEndian

BigEndian

data AlignmentInfo

An AlignmentInfo describes how a given type must and would best be aligned

Constructors

AlignmentInfo

Fields

abiAlignment :: Word32

preferredAlignment :: Maybe Word32

data AlignType

A type of type for which AlignmentInfo may be specified

Constructors

IntegerAlign

VectorAlign

FloatAlign

AggregateAlign

StackAlign

data DataLayout

a description of the various data layout properties which may be used during optimization

Constructors

DataLayout

Fields

endianness :: Maybe Endianness

stackAlignment :: Maybe Word32

pointerLayouts :: Map AddrSpace (Word32, AlignmentInfo)

typeLayouts :: Map (AlignType, Word32) AlignmentInfo

nativeSizes :: Maybe (Set Word32)

defaultDataLayout :: DataLayout

a DataLayout which specifies nothing