LLVM.General.AST.Global

Description

Globals - top-level values in Modules - and supporting structures.

Documentation

data Global

http://llvm.org/doxygen/classllvm_1_1GlobalValue.html

Constructors

GlobalVariable
http://llvm.org/docs/LangRef.html#global-variables

Fields

name :: Name

linkage :: Linkage

visibility :: Visibility

isThreadLocal :: Bool

addrSpace :: AddrSpace

hasUnnamedAddr :: Bool

isConstant :: Bool

type' :: Type

initializer :: Maybe Constant

section :: Maybe String

alignment :: Word32

GlobalAlias
http://llvm.org/docs/LangRef.html#aliases

Fields

name :: Name

linkage :: Linkage

visibility :: Visibility

type' :: Type

aliasee :: Constant

Function
http://llvm.org/docs/LangRef.html#functions

Fields

linkage :: Linkage

visibility :: Visibility

callingConvention :: CallingConvention

returnAttributes :: [ParameterAttribute]

returnType :: Type

name :: Name

parameters :: ([Parameter], Bool)

snd indicates varargs

functionAttributes :: [FunctionAttribute]

section :: Maybe String

alignment :: Word32

garbageCollectorName :: Maybe String

basicBlocks :: [BasicBlock]

data Parameter

Parameters for Functions

Constructors

Parameter Type Name [ParameterAttribute]

data BasicBlock

http://llvm.org/doxygen/classllvm_1_1BasicBlock.html LLVM code in a function is a sequence of BasicBlocks each with a label, some instructions, and a terminator.

Constructors

BasicBlock Name [Named Instruction] (Named Terminator)

globalVariableDefaults :: Global

helper for making GlobalVariables

globalAliasDefaults :: Global

helper for making GlobalAliass

functionDefaults :: Global

helper for making Functions