LLVM.General.AST.Operand

Description

A type to represent operands to LLVM Instructions

Documentation

newtype MetadataNodeID

A MetadataNodeID is a number for identifying a metadata node. Note this is different from named metadata, which are represented with NamedMetadataDefinition.

Constructors

MetadataNodeID Word

data MetadataNode

http://llvm.org/docs/LangRef.html#metadata

Constructors

MetadataNode [Maybe Operand]

MetadataNodeReference MetadataNodeID

data Operand

An Operand is roughly that which is an argument to an Instruction

Constructors

LocalReference Name
%foo

ConstantOperand Constant
Constants include GlobalReference, for @foo

MetadataStringOperand String

MetadataNodeOperand MetadataNode

type CallableOperand = Either InlineAssembly Operand

The Call instruction is special: the callee can be inline assembly