LLVM.General.AST.Type

Description

A representation of an LLVM type

Documentation

data FloatingPointFormat

LLVM supports some special formats floating point format. This type is to distinguish those format. I believe it's treated as a format for a float, as opposed to a vector of two floats, because its intended usage is to represent a single number with a combined significand.

Constructors

IEEE

DoubleExtended

PairOfFloats

data Type

http://llvm.org/docs/LangRef.html#type-system

Constructors

VoidType
http://llvm.org/docs/LangRef.html#void-type

IntegerType
http://llvm.org/docs/LangRef.html#integer-type

Fields

typeBits :: Word32

PointerType
http://llvm.org/docs/LangRef.html#pointer-type

Fields

pointerReferent :: Type

pointerAddrSpace :: AddrSpace

FloatingPointType
http://llvm.org/docs/LangRef.html#floating-point-types

Fields

typeBits :: Word32

floatingPointFormat :: FloatingPointFormat

FunctionType
http://llvm.org/docs/LangRef.html#function-type

Fields

resultType :: Type

argumentTypes :: [Type]

isVarArg :: Bool

VectorType
http://llvm.org/docs/LangRef.html#vector-type

Fields

nVectorElements :: Word32

elementType :: Type

StructureType
http://llvm.org/docs/LangRef.html#structure-type

Fields

isPacked :: Bool

elementTypes :: [Type]

ArrayType
http://llvm.org/docs/LangRef.html#array-type

Fields

nArrayElements :: Word64

elementType :: Type

NamedTypeReference Name
http://llvm.org/docs/LangRef.html#opaque-structure-types

MetadataType
http://llvm.org/docs/LangRef.html#metadata-type